more typescript, fix test
