+++ package.json
@@ -1,7 +1,9 @@
 {
   "scripts": {
     "lint": "tslint --fix --config ./tslint.json --project ./ --format stylish",
-    "compile": "npm run-script compile:grammar && npm run-script compile:ts",
+    "compile": "npm run-script compile:grammar && npm run-script compile:browse-games",
+    "precompile:browse-games": "npm run-script compile:ts",
+    "compile:browse-games": "node ./lib/cli/buildGameIcons.js",
     "compile:grammar": "nearleyc --minify-rule-names ./src/parser/grammar.ne > ./src/parser/grammar.ts",
     "compile:ts": "tsc && cp ./src/sound/sfxr.js ./lib/sound/ && cp ./src/sound/sfxr-browser.js ./lib/sound/",
     "ci": "npm run-script test",
@@ -21,10 +23,11 @@
     "pretest": "npm run-script compile:grammar && node ./script/doBatchStuff.js",
     "test": "npm run-script test:unit && npm run-script test:browser",
     "test:unit": "VERIFY_MATCHES=true NODE_ENV=development jest --coverage --updateSnapshot --verbose --detectLeaks --detectOpenHandles --testPathPattern '^((?!browser).)*\\.spec\\.ts'",
-    "pretest:browser": "npm run-script build",
+    "pretest:browser": "npm run-script build && npm run-script compile:browse-games",
     "test:browser": "NODE_ENV=development jest --updateSnapshot --verbose --detectOpenHandles --testPathPattern 'browser\\.spec\\.ts'",
-    "posttest": "npm run-script compile:ts && npm run-script test:cli && npm run-script coverage && npm run-script lint",
+    "posttest": "npm run-script test:cli && npm run-script coverage && npm run-script lint",
     "test:debug": "NODE_ENV=development node --inspect-brk -r source-map-support/register $(npm bin)/jest -i --updateSnapshot --detectLeaks --detectOpenHandles",
+    "pretest:cli": "npm run-script compile:ts",
     "test:cli": "istanbul cover -x '**/node_modules/**' -x 'lib/sound/sfxr.js' ./bin/puzzlescript.js -- --no-ui --game 'Beam Islands' --level 1 --size small --resume",
     "posttest:cli": "remap-istanbul --exclude bin/puzzlescript.js --input ./coverage/coverage.json --output ./coverage/coverage-cli.json && rm ./coverage/coverage.json",
     "test:server": "serve --listen 8000 .",
