+++ typings/svg-to-png.d.ts
@@ -1,0 +1,7 @@
+declare module "svg-to-png" {
+    interface SvgToPng {
+        convert(input: string | string[], outputDir: string, options: any): void
+    }
+    var v: SvgToPng
+    export = v
+}
